+++ kythe/javatests/com/google/devtools/kythe/analyzers/java/testdata/pkg/Imports.java
@@ -2,6 +2,11 @@
 //- PkgPackage.node/kind package
 package pkg;
 
+//- @CASE_INSENSITIVE_ORDER ref StringConst
+//- @String ref String
+import static java.lang.String.CASE_INSENSITIVE_ORDER;
+// TODO(schroederc): handle overloaded static method imports
+
 //- @"java.util" ref UtilPackage
 //- UtilPackage.node/kind package
 //- @List ref ListI
@@ -11,5 +16,12 @@ import java.util.LinkedList;
 
 public class Imports {
   //- @List ref ListI
-  public static void forAll(List<String> lst) {}
+  public static void forAll(List<String> lst) {
+    //- @String ref String
+    //- @CASE_INSENSITIVE_ORDER ref StringConst
+    System.err.println(String.CASE_INSENSITIVE_ORDER);
+  }
 }
+
+//- String.node/kind record
+//- String named vname("java.lang.String","","","","java")